requirements-extras.txt
django-debug-toolbar==1.9.1
django-extensions==2.0.7
django-jet==1.0.7
djangorestframework==3.8.2
google-api-python-client==1.4.1
flake8
